+++ build_runner/lib/build_runner.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 export 'src/asset/file_based.dart';
-export 'src/asset/reader.dart' show RunnerAssetReader, DigestAssetReader;
+export 'src/asset/reader.dart' show RunnerAssetReader;
 export 'src/asset/writer.dart';
 export 'src/generate/build.dart';
 export 'src/generate/build_result.dart';